refactor: rename and relocate files part 4 (final) -  have two Buttons that need to be addressed/consolidated (ultimately, delete the button under `app/ui/common/button.tsx`)
