+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Lexend } from "next/font/google";
 import "./globals.css";
-import NavBar from "../components/NavBar/NavBar";
+import NavBar from "../components/NavBar";
 
 const lexend = Lexend({ subsets: ["latin"] });
