+++ src/components/common/icons/X/index.tsx
@@ -1,0 +1,19 @@
+const XIcon = () => {
+  return (
+    <svg
+      width="14"
+      height="14"
+      viewBox="0 0 22 22"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M2 2L11 11M20 20L11 11M11 11L20 2M11 11L2 20"
+        stroke="currentColor"
+        strokeWidth="4"
+      />
+    </svg>
+  );
+};
+
+export default XIcon;
